requirements.txt
psycopg2-binary==2.8.6
git+https://github.com/jina-ai/jina-commons
pytest-mock
hnswlib==0.5.2
bidict==0.21.3